refactor: 카카오맵 이벤트 처리 로직 분리
