+++ android/app/src/main/java/io/coursepick/coursepick/view/KakaoMapManager.kt
@@ -1,8 +1,6 @@
 package io.coursepick.coursepick.view
 
 import android.Manifest
-import android.graphics.Point
-import android.graphics.PointF
 import android.location.Location
 import android.widget.Toast
 import androidx.annotation.RequiresPermission
@@ -14,9 +12,6 @@ import io.coursepick.coursepick.R
 import io.coursepick.coursepick.domain.Coordinate
 import io.coursepick.coursepick.domain.Latitude
 import io.coursepick.coursepick.domain.Longitude
-import io.coursepick.coursepick.domain.Segment
-import kotlin.math.pow
-import kotlin.math.sqrt
 
 class KakaoMapManager(
     private val mapView: MapView,
@@ -25,6 +20,7 @@ class KakaoMapManager(
     private val lifecycleHandler = KakaoMapLifecycleHandler(mapView)
     private val cameraController = KakaoMapCameraController(mapView.context)
     private val drawer = KakaoMapDrawer(mapView.context)
+    private val eventHandler = KakaoMapEventHandler()
     private var kakaoMap: KakaoMap? = null
 
     val cameraPosition get(): LatLng? = kakaoMap?.cameraPosition?.position
@@ -90,19 +86,18 @@ class KakaoMapManager(
         courses: List<CourseItem>,
         onClick: (CourseItem) -> Unit,
     ) {
-        kakaoMap?.setOnMapClickListener { map: KakaoMap, _, target: PointF, _ ->
-            courses.forEach { course: CourseItem ->
-                if (course.isNear(map, target)) {
-                    onClick(course)
-                    return@forEach
-                }
+        kakaoMap?.let { map: KakaoMap ->
+            eventHandler.setOnCourseClickListener(map, courses) { course: CourseItem ->
+                onClick(course)
             }
         }
     }
 
     fun setOnCameraMoveListener(onCameraMove: () -> Unit) {
-        kakaoMap?.setOnCameraMoveStartListener { _, _ ->
-            onCameraMove()
+        kakaoMap?.let { map: KakaoMap ->
+            eventHandler.setOnCameraMoveListener(map) {
+                onCameraMove()
+            }
         }
     }
 
@@ -157,25 +152,5 @@ class KakaoMapManager(
             },
             onFailure = onFailure,
         )
-    }
-
-    private fun CourseItem.isNear(
-        map: KakaoMap,
-        target: PointF,
-    ): Boolean {
-        val points: List<Point?> =
-            segments.flatMap(Segment::coordinates).map { coordinate: Coordinate ->
-                map.toScreenPoint(coordinate.toLatLng())
-            }
-        return (points.any { point: Point? -> point != null && point.isNear(target) })
-    }
-
-    private fun Point.isNear(point: PointF): Boolean {
-        val distance = sqrt((this.x - point.x).pow(2) + (this.y - point.y).pow(2))
-        return distance <= NEAR_TOUCH_THRESHOLD
-    }
-
-    companion object {
-        private const val NEAR_TOUCH_THRESHOLD = 50
     }
 }
